import limits, fix to search, versioning restart
Added some import limits to avoid severe compute/memory usage.  Replaced search with the correct file.  Moved version history <1.0 to a text file and 'restarted' that counter.  Some minor fixes to error handling and presentation tags (still out of date, but shouldnt be an issue for anyone but me, haha)